Split pycolmap into multiple compilation units
