+++ src/pycolmap/estimators/bindings.cc
@@ -1,5 +1,3 @@
-#pragma once
-
 #include "pycolmap/estimators/absolute_pose.h"
 #include "pycolmap/estimators/alignment.h"
 #include "pycolmap/estimators/cost_functions.h"
