+++ src/pycolmap/pipeline/extract_features.h
@@ -13,7 +13,6 @@
 
 #include <memory>
 
-#include <pybind11/iostream.h>
 #include <pybind11/pybind11.h>
 #include <pybind11/stl.h>
 #include <pybind11/stl_bind.h>
