+++ src/pycolmap/pipeline/images.h
@@ -15,7 +15,6 @@
 #include <memory>
 
 #include <glog/logging.h>
-#include <pybind11/iostream.h>
 #include <pybind11/pybind11.h>
 #include <pybind11/stl.h>
 #include <pybind11/stl_bind.h>
