+++ src/pycolmap/pipeline/mvs.h
@@ -12,7 +12,6 @@
 
 #include "pycolmap/helpers.h"
 
-#include <pybind11/iostream.h>
 #include <pybind11/pybind11.h>
 #include <pybind11/stl.h>
 #include <pybind11/stl_bind.h>
